move patchFlowVariables into ProblemSpecificFunctions
