+++ CMake/BuildPeleExe.cmake
@@ -46,8 +46,6 @@ function(build_pele_exe pele_exe_name pele_physics_lib_name)
        ${SRC_DIR}/PeleLMeX_Forces.cpp
        ${SRC_DIR}/PeleLMeX_Init.cpp
        ${SRC_DIR}/PeleLMeX_ODEQty.cpp
-       ${SRC_DIR}/PeleLMeX_PatchFlowVariables.H
-       ${SRC_DIR}/PeleLMeX_PatchFlowVariables.cpp
        ${SRC_DIR}/PeleLMeX_Plot.cpp
        ${SRC_DIR}/PeleLMeX_Projection.cpp
        ${SRC_DIR}/PeleLMeX_Reactions.cpp
