+++ Exec/RegTests/EB_BackwardStepFlame/GNUmakefile
@@ -25,9 +25,6 @@ DEBUG = FALSE
 FSANITIZER = FALSE
 THREAD_SANITIZER = FALSE
 
-# PeleLMeX
-CEXE_sources += PeleLMeX_PatchFlowVariables.cpp
-
 # PelePhysics
 Chemistry_Model = drm19
 Eos_Model = Fuego
