+++ Source/Make.package
@@ -8,7 +8,6 @@ CEXE_headers += PeleLMeX_DiffusionOp.H
 CEXE_headers += PeleLMeX_DeriveFunc.H
 CEXE_headers += PeleLMeX_FlowControllerData.H
 CEXE_headers += PeleLMeX_BPatch.H
-CEXE_headers += PeleLMeX_PatchFlowVariables.H
 
 ## Sources
 CEXE_sources += main.cpp
@@ -43,7 +42,6 @@ CEXE_sources += PeleLMeX_Diagnostics.cpp
 CEXE_sources += PeleLMeX_FlowController.cpp
 CEXE_sources += PeleLMeX_DeriveUserDefined.cpp
 CEXE_sources += PeleLMeX_BPatch.cpp
-CEXE_sources += PeleLMeX_PatchFlowVariables.cpp
 
 ifeq ($(USE_SOOT), TRUE)
   CEXE_sources += PeleLMeX_Soot.cpp
